pyproject.toml
[project]
name = "comparison"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[all]>=0.116.1",
    "reportlab>=4.4.2",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["data*", "docs*", "tests*"]
